requirements.txt
# Core Azure and LLM packages
openai>=1.35.0
azure-ai-ml>=1.15.0
azure-identity>=1.15.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Environment and utilities
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.31.0

# Meta/Facebook integration
facebook-business>=19.0.0

# Google Cloud (for BigQuery)
google-cloud-bigquery>=3.11.0

# Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: Image processing
opencv-python>=4.8.0

# Optional: Advanced LLM frameworks
langchain>=0.1.0
langchain-openai>=0.1.0

# Testing and development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0